+++ docs/changelog.md
@@ -18,13 +18,12 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 ### TODO
 
 - Check out how many time render is invoked, and optimize it as done with Selector.
-- Add @flow to Canvas and FlowViewFrame.
-- Pass theme as argument to Canvas constructor, create a dark (or whatever) theme example.
 
 ## Added
 
 - Trigger *updateNodesGeometry* event, when nodes are moved.
 - Complete flowjs coverage.
+- Pass theme as argument to Canvas constructor, created a dark theme example.
 
 ## [v2.18.0] - 2017-11-23